Рефакторинг. Вынес всё связанное с трекером задач в отдельный пакет
