+++ lesson_5/src/main/java/name/imh0t3mp/course/geekbrains/task_tracker/TaskStatus.java
@@ -1,4 +1,4 @@
-package name.imh0t3mp.course.geekbrains;
+package name.imh0t3mp.course.geekbrains.task_tracker;
 
 /**
  * Перечень возможных статусов задачи
