+++ lesson_5/src/main/java/name/imh0t3mp/course/geekbrains/task_tracker/errors/MyArrayDataException.java
@@ -1,4 +1,4 @@
-package name.imh0t3mp.course.geekbrains.errors;
+package name.imh0t3mp.course.geekbrains.task_tracker.errors;
 
 public class MyArrayDataException extends Exception {
     public MyArrayDataException(String message) {
