+++ lesson_5/src/main/java/name/imh0t3mp/course/geekbrains/task_tracker/errors/TaskAlreadyExists.java
@@ -1,4 +1,4 @@
-package name.imh0t3mp.course.geekbrains.errors;
+package name.imh0t3mp.course.geekbrains.task_tracker.errors;
 
 public class TaskAlreadyExists extends Exception {
     public TaskAlreadyExists(String s) {
